requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.11.4
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.24
langchain-text-splitters==0.2.4
openai
pandas
pillow
pypdf
python-dotenv
supabase
fastapi
uvicorn
python-multipart
opencv-python-headless
reportlab
openpyxl
langchain-google-genai<2.0.0
zstandard
numpy<2
